Feat prepend method
It adds a new value to the start of the linked list, as opposed to appe-
nd which adds at the end of it. The new node's value is given by the pa-
rameter of the method. If the list is empty, head and tail both become
the new node.